Timer is mostly functional and in the right spot. Can't get the stopTimer function working yet.
